requirements_txt.txt
python-telegram-bot==13.7
requests==2.31.0
urllib3
Flask==2.3.3
gunicorn==21.2.0